restructure and make lint
