+++ .tool/lint
@@ -14,12 +14,11 @@ for d in $(find . -type d -not -iwholename '*.git*' -a -not -iname '.tool' -a -n
 		--exclude='error return value not checked.*(Close|Log|Print).*\(errcheck\)$' \
 		--exclude='.*_test\.go:.*error return value not checked.*\(errcheck\)$' \
 		--exclude='duplicate of.*_test.go.*\(dupl\)$' \
-		--exclude='schema/fs.go' \
 		--exclude='duplicate of.*main.go.*\(dupl\)$' \
 		--disable=aligncheck \
 		--disable=gotype \
 		--disable=gas \
-		--cyclo-over=35 \
+		--cyclo-over=50 \
 		--tests \
 		--deadline=10s "${d}"
 done
